feat: separated completions, added higherorder completion
