+++ src/lib/lang-miro/index.ts
@@ -17,7 +17,9 @@ import {
 } from '@codemirror/autocomplete';
 import { MiroHighlighting, MiroStyleTags } from './highlight';
 import { linter, Diagnostic } from '@codemirror/lint';
-import { LooseIcon, allIconNames } from '@/components/icons';
+import { allIconNames } from '@/components/icons';
+import { SyntaxNode } from '@lezer/common';
+import { iconCompletion, miroTypeCompletion } from './completions';
 
 let parserWithMetadata = parser.configure({
   props: [
@@ -43,46 +45,7 @@ export const miroLanguage = LRLanguage.define({
   parser: parserWithMetadata,
 });
 
-export const miroDefaultCompletion = completeFromList([
-  { label: 'string', type: 'type' },
-  { label: 'number', type: 'type' },
-  { label: 'boolean', type: 'type' },
-  { label: 'int', type: 'type' },
-  { label: 'float', type: 'type' },
-  { label: 'date', type: 'type' },
-]);
-
-export const iconCompletion = completeFromList(
-  allIconNames.map(icon => ({
-    label: `i-${icon}`,
-    type: 'icon',
-  })),
-);
-
-const miroTypeCompletion = (
-  context: CompletionContext,
-): CompletionResult | null => {
-  // match if there is a word and space
-  const match = context.matchBefore(/\w\s+\w+/);
-
-  if (match) {
-    return {
-      from: context.pos,
-      options: [
-        { label: 'string', type: 'type' },
-        { label: 'number', type: 'type' },
-        { label: 'boolean', type: 'type' },
-        { label: 'int', type: 'type' },
-        { label: 'float', type: 'type' },
-        { label: 'date', type: 'type' },
-      ],
-    };
-  }
-
-  return null;
-};
-
-const MiroLinter = linter(view => {
+export const MiroLinter = linter(view => {
   let diagnostics: Diagnostic[] = [];
   syntaxTree(view.state)
     .cursor()
@@ -93,14 +56,6 @@ const MiroLinter = linter(view => {
           to: node.to,
           severity: 'error',
           message: 'Syntax Error!',
-          actions: [
-            {
-              name: 'Remove',
-              apply(view, from, to) {
-                view.dispatch({ changes: { from, to } });
-              },
-            },
-          ],
         });
     });
   return diagnostics;
@@ -111,7 +66,7 @@ export function miro() {
     MiroHighlighting,
     MiroLinter,
     autocompletion({
-      override: [miroTypeCompletion, completeAnyWord, iconCompletion],
+      override: [miroTypeCompletion, iconCompletion],
     }),
   ]);
 }
